refactor(indexer-alt): move watermark types
## Description

Move `*Watermark` types under `models`, as that's where they were before
the introduction of a framework.

## Test plan

CI
