+++ crates/sui-indexer-alt-framework/src/models/mod.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod cp_sequence_numbers;
+pub(crate) mod watermarks;
