+++ crates/sui-indexer-alt-framework/src/pipeline/concurrent/commit_watermark.rs
@@ -18,8 +18,8 @@ use tracing::{debug, error, info, warn};
 
 use crate::{
     metrics::{CheckpointLagMetricReporter, IndexerMetrics},
+    models::watermarks::CommitterWatermark,
     pipeline::{logging::WatermarkLogger, CommitterConfig, WatermarkPart, WARN_PENDING_WATERMARKS},
-    watermarks::CommitterWatermark,
 };
 
 use super::Handler;
