+++ crates/sui-indexer-alt-framework/src/pipeline/sequential/mod.rs
@@ -9,9 +9,9 @@ use sui_types::full_checkpoint_content::CheckpointData;
 use tokio::{sync::mpsc, task::JoinHandle};
 use tokio_util::sync::CancellationToken;
 
-use crate::{metrics::IndexerMetrics, watermarks::CommitterWatermark};
-
 use super::{processor::processor, CommitterConfig, Processor, PIPELINE_BUFFER};
+
+use crate::{metrics::IndexerMetrics, models::watermarks::CommitterWatermark};
 
 use self::committer::committer;
